refactor: Move solvers to separate subdirectories
